LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/drawing-text-to-the-canvas/2d.text.drawing.style.fontVariant.settings-expected.txt

2d.text.drawing.style.fontVariant.settings
Testing basic functionalities of fontKerning for canvas
Actual output:

FAIL Testing basic functionalities of fontKerning for canvas assert_equals: ctx.fontVariantCaps === "Normal" (got [undefined], expected Normal[string]) expected (string) "Normal" but got (undefined) undefined
